crawler/requirements.txt
# Web Scraping
selenium==4.18.1
beautifulsoup4==4.12.3
webdriver-manager==4.0.1

# HTTP Requests
requests==2.31.0

# Database
psycopg2-binary==2.9.9

# Environment variables
python-dotenv==1.0.1

# Data processing
pandas==2.2.1
